+++ modules/combineRules.js
@@ -1,5 +1,7 @@
 import assign from './utils/assign'
 
 export default function combineRules(...rules) {
-  return props => rules.reduce((style, rule) => assign(style, rule(props)), { })
+  return function combined(props) {
+    return rules.reduce((style, rule) => assign(style, rule(props)), { })
+  }
 }