+++ modules/createRenderer.js
@@ -50,7 +50,7 @@ export default function createRenderer(config = { }) {
      * @param {Object?} props - properties used to render
      * @return {string} className to reference the rendered rule
      */
-    renderRule(rule, props = { }) {
+    renderRule(rule, props = { }, _selectorPrefix = '') {
       // rendering a rule for the first time
       // will create an ID reference
       if (renderer.ids.indexOf(rule) < 0) {
@@ -59,7 +59,7 @@ export default function createRenderer(config = { }) {
         // directly render the static base style to be able
         // to diff future dynamic style with those
         if (Object.keys(props).length > 0) {
-          renderer.renderRule(rule, { })
+          renderer.renderRule(rule, { }, _selectorPrefix)
         }
       }
 
@@ -67,7 +67,7 @@ export default function createRenderer(config = { }) {
       const ruleId = renderer.ids.indexOf(rule)
 
       const classNamePrefix = renderer.prettySelectors && rule.name ? rule.name + '_' : 'c'
-      const className = classNamePrefix + ruleId + generatePropsReference(props)
+      const className = _selectorPrefix + classNamePrefix + ruleId + generatePropsReference(props)
 
       // only if the cached rule has not already been rendered
       // with a specific set of properties it actually renders